style: moreinfopage comments improved
